Moved module into a folder
